+++ ChangeLog.md
@@ -7,6 +7,7 @@
 - Add method into the class `myDataset` to get the fields.
 - Add the class `myDatasetIterator` to travel a whole dataset.
 - Add the class `myNetwork` to predict.
+- Update the class `myNetwork` to calculate loss and gradient.
 
 ### Changed
 - Update the class `myDatasetIterator` for batch training.